+++ framework/Cli/Controllers/WebCliController.php
@@ -3,6 +3,7 @@
 namespace Framework\Cli\Controllers;
 
 use Framework\Cli\Cli;
+use Framework\Facades\Http;
 use Framework\Routing\BaseController;
 use Framework\Routing\ControllerInterface;
 
@@ -10,14 +11,15 @@ class WebCliController extends BaseController implements ControllerInterface
 {
     public function execute(): void
     {
-        if ($this->getRequestMethod() === 'GET') {
+        if (Http::requestMethod() === 'GET')
+        {
             view('framework.cli.webcli');
             return;
         }
 
-        if ($this->getRequestMethod() === 'POST') {
-            $cli = new Cli();
-            $cli->runWebCli($this->getParam('command', ''));
+        if (Http::requestMethod() === 'POST')
+        {
+            Cli::runWebCli(Http::param('command', ''));
             return;
         }
     }